# OEMBackend/src/api/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"TaskCategoryDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"category",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResourceDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"startTime": {
						"type": "string"
					},
					"endTime": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OperationDto": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/TaskCategoryDto"
					},
					"startTime": {
						"type": "string"
					},
					"endTime": {
						"type": "string"
					},
					"resources": {
						"items": {
							"$ref": "#/components/schemas/ResourceDto"
						},
						"type": "array"
					},
					"payload": {}
				},
				"required": [
					"type",
					"startTime",
					"endTime",
					"resources"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreateTaskCategoryDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"category",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Page_TaskCategoryDto_": {
				"properties": {
					"pageNumber": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"pageCount": {
						"type": "number",
						"format": "double"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/TaskCategoryDto"
						},
						"type": "array"
					}
				},
				"required": [
					"pageNumber",
					"pageSize",
					"pageCount",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WorkQueueItem": {
				"properties": {
					"day": {
						"type": "string"
					},
					"alg": {
						"type": "string"
					},
					"daysAhead": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"day",
					"alg",
					"daysAhead"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ScheduleQueueItem": {
				"properties": {
					"id": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/WorkQueueItem"
					},
					"priority": {
						"type": "number",
						"format": "double"
					},
					"result": {},
					"requestedAt": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"type": "string"
					},
					"issuer": {
						"type": "string"
					},
					"estimatedStartTime": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"estimatedEndTime": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"id",
					"data",
					"priority",
					"requestedAt",
					"status",
					"issuer",
					"estimatedStartTime",
					"estimatedEndTime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OperationPlanMetadataDto": {
				"properties": {
					"createdBy": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"algorithmUsed": {
						"type": "string"
					}
				},
				"required": [
					"createdBy",
					"createdAt",
					"algorithmUsed"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OperationPlanDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"relatedVVN": {
						"type": "string"
					},
					"dock": {
						"type": "string"
					},
					"operationSchedule": {
						"items": {
							"$ref": "#/components/schemas/OperationDto"
						},
						"type": "array"
					},
					"metadata": {
						"$ref": "#/components/schemas/OperationPlanMetadataDto"
					}
				},
				"required": [
					"id",
					"relatedVVN",
					"dock",
					"operationSchedule",
					"metadata"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Page_OperationPlanDto_": {
				"properties": {
					"pageNumber": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"pageCount": {
						"type": "number",
						"format": "double"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/OperationPlanDto"
						},
						"type": "array"
					}
				},
				"required": [
					"pageNumber",
					"pageSize",
					"pageCount",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IncidentTypeDto": {
				"properties": {
					"bid": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"severity": {
						"type": "string"
					},
					"subtypeOf": {
						"type": "string"
					},
					"subtypes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"bid",
					"name",
					"description",
					"severity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Page_IncidentTypeDto_": {
				"properties": {
					"pageNumber": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"pageCount": {
						"type": "number",
						"format": "double"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/IncidentTypeDto"
						},
						"type": "array"
					}
				},
				"required": [
					"pageNumber",
					"pageSize",
					"pageCount",
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PartialIncidentTypeDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"severity": {
						"type": "string"
					},
					"subtypeOf": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"description",
					"severity"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "oembackend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/task-categories": {
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TaskCategoryDto"
								}
							}
						}
					}
				},
				"tags": [
					"Task Categories"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTaskCategoryDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Page_TaskCategoryDto_"
								}
							}
						}
					}
				},
				"tags": [
					"Task Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "pageNumber",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/task-categories/{id}": {
			"put": {
				"operationId": "UpdateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/TaskCategoryDto"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Task Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTaskCategoryDto"
							}
						}
					}
				}
			}
		},
		"/task-categories/code/{code}": {
			"get": {
				"operationId": "GetCategoryByCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/TaskCategoryDto"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Task Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/schedule/request": {
			"get": {
				"operationId": "ScheduleRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Scheduling"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "day",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "alg",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "daysAhead",
						"required": false,
						"schema": {
							"default": 2,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/schedule/acceptRequest": {
			"post": {
				"operationId": "AcceptRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ScheduleQueueItem"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Scheduling"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/schedule/rejectRequest": {
			"post": {
				"operationId": "RejectRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ScheduleQueueItem"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Scheduling"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/operation-plans": {
			"get": {
				"operationId": "GetPlans",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Page_OperationPlanDto_"
								}
							}
						}
					}
				},
				"tags": [
					"Operation Plans"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "pageNumber",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "startDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/operation-plans/{id}": {
			"get": {
				"operationId": "GetPlanById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/OperationPlanDto"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Operation Plans"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateOperationPlan",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/OperationPlanDto"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Operation Plans"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
...
		"/operation-plans/regenerate": {
			"post": {
				"operationId": "RegeneratePlansForDay",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Operation Plans"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"daysAhead": {
										"type": "number",
										"format": "double"
									},
									"algorithm": {
										"type": "string"
									},
									"day": {
										"type": "string"
									}
								},
								"required": [
									"algorithm",
									"day"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/incident-types": {
			"get": {
				"operationId": "GetPaged",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Page_IncidentTypeDto_"
								}
							}
						}
					}
				},
				"tags": [
					"Incident Types"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "pageNumber",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateIncidentType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IncidentTypeDto"
								}
							}
						}
					}
				},
				"tags": [
					"Incident Types"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PartialIncidentTypeDto"
							}
						}
					}
				}
			}
		},
		"/incident-types/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IncidentTypeDto"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Incident Types"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateIncidentType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IncidentTypeDto"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Incident Types"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PartialIncidentTypeDto"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}